mysql: Migrate away from long-deprecated mysql module to mysqli
The mysql module for php has been deprecated for years (PHP5.5).
Even by that time, it was highly discourage to use since PHP5.0/
MySQL 4.1. PHP7.X finally gets rid of it. I am surprised that a
relatively recent project chose it as a module.

Replace it, which is relatively strightforward, so it can run
on newer php versions.

Bug: T224589
Change-Id: Iac53c01b01d4e7483212e1b67540ce388e8bf433